moved duplicate code to a separate class
